requirements.txt
pyyaml~=5.3
docopt~=0.6